Spitting (WIP, Buggy)
They split, but sometimes has erratic behavior such as splitting into two identical sizes of the size of the original asteroid rather than splitting into the smaller childs.

Note: the player script in this commit has being destroyed by asteroids commented out, due to testing on PC and not being able to move it was hard for me to test without dying a lot.